Move page components into separate folder
